Fix check for implemented interfaces (dart-lang/source_gen#111)
The `interfaces` getter does not behave as hoped. See
https://github.com/dart-lang/sdk/issues/27947

This code was safe before since List happened to explicitly have
`implements Iterable`, but now that there is an
`EfficientLengthIterable` interface in between it isn't working.

Recursively check interfaces the same way we recursively check the
superclass. This now ignores the edge case where multiple interfaces
matched the tester and will return the first match.