Fix race condition with delete events for generated files, fixes https://github.com/dart-lang/build/issues/175 (#185)
As a part of a normal build, invalidated outputs are deleted at the start, causing a `delete` event to be captured by the watcher. This ensures that those delete events are not tracked as real edits, and don't cause an infinite loop of builds.

In practice on linux the add events for the files being recreated were coming through fast enough that we were ignoring the delete events anyways, but on mac it was consistently not happening that way (the add event came through to late for some files, after we started the next build).