Fix Analyzer warning in latest dev SDK
This should fix the failure on Travis.

See https://github.com/dart-lang/sdk/issues/29164

Since the latest analyzer in strong mode uses downward inference from
the return type it expects the type parameter T on the `fold` call to be
`FutureOr<List<int>>` and so the callback we are passign does not fit
the signature it expects. Work around this by specifying the type
parameter ourselves rather than havt it get inferred.