Always check BuildStatus after builds
It is a lot easier to track failures down if we check the status before
the outputs since otherwise a failing build hides it's exception and
instead prints an error about a missing output.

Also run serve and watch tests with deleteFilesByDefault - this was
added to testPhases already but these tests go aroudn that.