Delete the _inputsByPackage variable
Moves all reads of this variable to reads of the (at the time) static
asset graph. With further refactoring the lines of communication can
become even more explicit with the AssetGraph rather than implicit and
order-dependent as they are with this variable.

There are very likely performance regressions with this change since
some Sets are built over and over from scratch rather than stored or
built up over time - we can address these later.