Remove some unnecessary checks
We don't need to check for duplicate adds - only the AssetGraph class
itself can add nodes and it can maintain the invariants it needs. When a
duplicate could have come up we were already doing a remove anyway.

We don't need to check for outputting over an input since checkd for
existing generated files after building up the graph.