Delete the _inputsByPackage variable (#330)
Moves all reads of this variable to reads of the (at the time) static
asset graph. With further refactoring the lines of communication can
become even more explicit with the AssetGraph rather than implicit and
order-dependent as they are with this variable.

There are very likely performance regressions with this change since
some Sets are built over and over from scratch rather than stored or
built up over time - we can address these later.

- Remove some unnecessary checks

We don't need to check for duplicate adds - only the AssetGraph class
itself can add nodes and it can maintain the invariants it needs. When a
duplicate could have come up we were already doing a remove anyway.

We don't need to check for outputting over an input since checked for
existing generated files after building up the graph.