Add an `action` argument to `resolveSource`
Closes #355

After the Resolver apis all became async it became impossible to do any
work with the Resolver instance without also passing a `tearDown`
argument. By default expect the work to be done now inside an `action`
callback and continue to allow the old behavior by returning the
resolver in the action and passing a teardown.

Change `tearDown` from a `Future<Null>` to a `Future` for a more
flexible api. Can become `Future<void>` once the language allows it.