Disallow reads outside the asset graph
Fixes #359

We used to rely on the graph being updated if it didn't already have the
source node but now expect the graph to keep it's structure after
initialization. In order to allow reads to any file that would have been
visible before add all packages as sources, not just those that have a
build action. This should only increase the size of the serialized graph
for projects which did not read these files anyway - which excludes
anything using BarbackResolvers and crawling imports.