Add `mapAssetIds` argument to `checkOutputs` (#377)
In order to accomplish #216 we will need `build_runner` to be able to
write assets to a different location on disk than their logical location
within a package. The translation will happen at a lower level than the
writer that gets spied on during the build so that apparent outputs from
the builder might not match the written assets to the
`InMemoryAssetWriter`. Add an argument that can perform the same mapping
as would happen in the actual asset writer during the build.

- Add a Doc comment to checkOutputs
- Add the mapAssetIds argument and default to a passthrough.
- Add tests for checkOutputs