Remove restrictions on outputs by package (#379)
Towards #378

We know that outputs can only be for the same package as inputs and we
check that the only assets written are declared outputs. We don't need
to also pass around a root package or force that all inputs are for the
same package - the build systems which call `runBuilder` will need to do
appropriate checks regardless of whether we duplicate them here.