Separate asset graph loading from build_impl
When we add support for writing to a cache directory the location of an
asset read by the AssetReader can vary based on whether it is a
GeneratedAssetNode so there may be a feedback loop between the
AssetGraph and the AssetReader. This is trick before the refactoring
since the AssetGraph can become null and rebuild over the course of the
build. Pull out the logic for loading the AssetGraph and BuildType into
a separate class so the lifecycle is explicit and contained.

- Add BuildDefinition and BuildDefinitionLoader extracted from BuildImpl
  to separate responsibilities.
- Refactor such that the two entry points - first build and subsequent
  builds - have their own logic rather than being interleaved.
- Remove explicit `_isFirstBuild` field since it can be inferred based
  on whether there are updates or an existing asset graph.
- Drop updateForSources method in AssetGraph. Instead, find the changes
  in the same format as `updates` and apply them consistently. Add
  support for new sources in `updateAndInvalidate`.

No new tests because no behavior has changed - I will add test of the
BuildDefinitionLoader when we need to change it's behavior.