Use a StreamQueue rather than a busy loop
The old pattern of checking for a new result every 10 milliseconds is
error prone - this is what `StreamQueue` was meant for.

We do lose some possibility to catch multiple builds that happened when
we only expected one - but even those would only be caught if they they
happened within 10ms of eachother.