Support running builders against other packages (#386)
Closes #216

Add an argument to `build` which indicates that generated files should
be written to `.dart_tool/build/generated/<package>/<path>` rather than
in the source directory for the root package. With this argument enabled
all writes, including for the root package, go to the generated
directory. If we want to support running 'local' and 'non-local'
builders in the same build we will need to make this more fine grained.

- Don't surface the `FileSystemException` when trying to glob a
  directory that doesn't exist.
- Add BuildCacheReader and BuildCacheWriter which translates paths for
  generated assets into the special directory.
- Don't enforce that builders run only on the root package when outputs
  are written to the generated cache.
- Plumb the `writeToCache` argument through to `build_impl` for both
  `build` and `watch`. We will need further changes in `serve` to
  support this (#384)