Refactor checks for file changes in between builds
Fixes #408

Previous approach would not see generated files in the asset cache since
it was using the reader from the options instead of the wrapped reader.
Rather than make the `BuildCacheReader` implement `RunnerAssetReader`
refactor the check for updates to use the already loaded
`currentSources` which does honor the asset cache.