add a basic graph inspector tool to build_runner (#441)
got a quick start on this - today its interactive but we might want to change that since you don't get much bash niceness but its definitely useful for inspecting graphs and provides a good extension point for future features.

Today this has two commands, inspect (for inspecting individual nodes), and graph (for inspecting the whole graph). We can rename these later too if we want.