Add back basic file caching (#530)
Fixes #526, well enough at least (It guarantees that all builders see the same version of a file within a build).

We could still end up scheduling an extra build that isn't necessary (if the file was first read during a build after the change happened), but that is less of an issue and could be resolved using asset hashes.

This also has the nice benefit of speeding up builds, especially the larger builds. Initial build for angular_components_example goes from ~140s for me to ~83s.