Deprecate inputs and excludes from PackageNode
The only use case for these was that we were depending on them being set
appropriately for the root and non-root packages. We had previously been
handling this in two places and it was changed to only be configued in
the PackageNode. Adding it back to BuildDefinition will allow us to
deprecate and remove the fields in the PackageGraph.