Change BuildApplication to user builder factories (#631)
Fixes #625

We'll eventually need to support overriding what the `BuildApplication`
says with per-package or per-module config, including passing different
arguments to each `Builder` instance. We need to pass through the
`BuilderFactory` functions and add the necessary keys to relate a config
with the builder.

In order to support the old use case with less noise add `applyToRoot`
with some hacks to plumb through support for checking which package node
is the root. Once we add an `isRoot` field to `PackageNode` we can solve
this with a normal `PackageFilter`.

For now this will end up constructing more `Builder` instances than is
necessary, we can add caching by unique sets of arguments as an
enhancement.