Add required_inputs config option (#640)
Towards #639

We want to be able to specify that `DDC` needs all builders which
generate Dart files to have already run before it can run. We won't have
a full order solver, but this will us allow us to solve the two use
cases we have seen so far: all `.dart` before a compiler, and all `.css`
before the Angular stylesheet compiler.