Skip translation to build cache for missing (#648)
While working towards #647 I needed to set up the reader such that it is
always wrapped to translate some assets to the generated asset
directory. We have a few tests which set up coniditions such that the
AssetGraph is empty, and they pass because they rely on not using a
BuildCacheReader. This change only is necessary to support those tests,
but it is safe in general because the delegate reader will handle
missing files anyway.