Add InMemoryAssetReader.shareAssetCache constructor
When we started copying asset rather than continuing to use the same Map
instance it broke an undocumented behavior that we relied on in some
tests - that we could automatically keep new written assets reflected by
the writer.

Keep the strong mode types accurate and add a new contstructor which
solves this use case specificall.