Add magic placeholder assets under lib, web, and test (#746)
These assets are available to use as primary inputs, but cannot be read (and will give an AssetNotFoundException if you do try to read them). They are represented by a new type of node, PlaceholderAssetNode.

Note that this is a potentially breaking change for builders which have over-broad input extensions (looking at you CopyBuilder). These builders may attempt to read the input and throw an exception, but I don't think this is a major issue since it should be a very uncommon case. It may however break existing tests for people using CopyBuilder (it broke a bunch of ours).

I also added a bunch of boolean getters to describe functionality of the different asset node types, and replaced is checks where possible with those.

Closes #710