Allow overriding `build.yaml` for other packages
While we're rolling this out, and likely into the future, it's handy to
be able to hack on another package's `build.yaml` without a path
override to a local copy of the package. Read files matching
`$packageName.build.yaml` and use them instead of the `build.yaml`
coming from the package.

- Add `build_config_overrides.dart` with a function to look for override
  config files in the current directory and parse them into a map.
- When generating the build script and creating build actions, first
  look for overrides and use them over the package provided config.