Filter out non-hidden actions outside the root
Previously this would be an error, but we want to be able to handle
cases like a `build.yaml` specifying `enable: true` for a Builder that
writes to the source tree, but then also being used as a dependency.

This is a small change for manual build scripts where instead of a
helpful exception an action would just not run.

- Allow BuildTo.source for non-root auto_apply in `build_config`
- Default `BuildTo` to `cache` always rather than depending on the value
  of `AutoApply`. Anything that writes to the source tree should be
  explicitly called out.
- Default `hideOutput` for most BuilderApplications to true to match the
  defaultin of `BuildTo`