Use an Isolate rather than a subprocess (#837)
We can avoid the `--assume-tty` hack since the Isolate sees the same TTY
as the VM. Leave that flag in for now as an option for higher levels of
wrapping scripts.

Only generate the String for the build script in
`build_script_generate.dart`. Handle writing it to a file in the main
method. This will open up the possibility of skipping the write to disk
and using a `data:` URI if we can solve the problem of build script
updates. If we can do that we can add a separate command to write the
build script to disk.