Add support for `--config` overriding build.yaml (#882)
Closes #881

Plumb the flag `--config` through to where we find build config
overrides. If it's non-null look for the file `build.$configKey.yaml`
and use it instead of `build.yaml`.

Rename the `dart2js` build config for `e2e_example` to the new pattern
and use it rather than rewrite it as `build.yaml` during the test.