Filter inputs to actions based sources
Fixes #1076
Fixes a missed case for #1055

Instead of checking whether the direct input to an action matches the
target sources and the generate_for, we need to check whether the
_original_ source was in the target and the _current_ input matches the
generate_for.

We had this check twice, once while building up the asset graph, and
once while doing the build. Shift instead to trusting that the outputs
for a given phase have already done necessary filtering and while
building only check the lazy phases and whether the input was written.

- Add `outputsForPhase` to `AssetGraph`. This immediately gets to the
  filtered and valid files that may be generated, and their
  `primaryInput` can point to the assets we should be running on.
  Initial naive implementation can be replaced with a cached version in
  a followup.
- Don't implement `matches` on `BuildAction`. Instead separately track
  the two types of Asset filtering so we can compare against different
  ids.
- Drop the `_MultiMatcher` since there is no longer any use.
- Fix or remove tests depending on the old behavior of not running
  to-source Builders on hidden primary inputs.
- Add regression tests both directly on graph building as well as
  through the overall build.