Catch exceptions and mark builds as failures
Fixes #1207

Catch exceptions and use them to mark individual actions as failures
rather than allowing the exception to bubble up and stop the entire
build.

Move the handling for `--fail-on-severe` to whether or not we mark an
action as failing rather than whether or not we treat failed actions as
a build failure.