Avoid `packagePath` use in `PackageReader`
Closes #1285

This method does not work when the `.packages` file does not have the
same semantics as the one created by pub - that is when it does use
point to directories named `lib/` under real packages. The package map
we create with `build_runner --output` does not have the same semantics,
it points to directoreis that are not naemd `lib/`.

- Implement `_resolve` on top of the package resolver directly resolving
  by URI for `package:` uris. This should always work.
- Implement `asset:` URIs by assuming the cwd is the root package, which
  is the only place we should get asset: URIs. This may be brittle but
  is the best we can do for now.
- Remove support for `findAssets` since the implementation is broken in
  these cases and it's not clear how we'd fix it.