Reimplement findAssets in PackageAssetReader (#1291)
This was removed because I (incorrectly) thought there were no usages
and it was hard to support. Add it back in with a hack of listing all
the files in the searched package, reconsituting a "lib/" structure, and
then running the glob against it.

This is potentially less performant for large packages and narrow
globs since we always list all files instead of just listing the glob.