Fix recording of input digests on the asset graph (#1320)
Depending on the order that inputs were read we could compute a
different `previousInputsDigest` than we would with identical inputs on
the next build. This would cause more actions to be run than necessary -
though it only impacts the case where all inputs were identical which is
relatively rare in practice, however upcoming changes to the module
strategy will make it much more likely.

- Store read assets in a sorted set which matches the way we store
inputs on the asset graph.
- Don't always include the primary input as if it was read - if the
build step doesn't read the primary input we can assume it does not
depend on rerunning when the content changes.
- Pull the digest computation out of the loop since we can always run it
against  the first output's node.