Cache only complete modules
Previous there was a race condition between updating `readMetas` and
populating `assetToPrimary` so we could attempt to pull out data that
was not ready yet and miss out on modules that should exist.

This changes the usage of `Resource` to include only data that will
never change, and includes access checks to make sure we only read data
that is allowed and tracked by the build system.