Add tools to generate test packages with builders
This aims to have a nicer workflow than the places where we have an
entire Dart file as a String in a test. It primarily relies on a trick
where it reades lines out of the _calling_ script and inserts them into
a test file.

A package can be generated with a `lib/builders.dart` and `build.yaml`,
or a package gan be generated with a `tool/build.dart`. In either case
the builders are wrapped automatically with factories and the remaining
configuration is plumbed through as appropriate.

There are limitations around what code can be imported and needed to
used only top-level variables rather than factories that look at
options. These limitations can be relaxed if the use case comes up.

Adds a `BuildTool` and `BuildServer` class to abstract the details of
interacting with the build system. Relies on the fact that the manual
script still uses `run` to have consistent interaction.