Handle spawn errors with an explicit error
Fixes #1221

Note that this doesn't fix the exact issue reported in Dart 1 runtime,
but does fix it in Dart 2 runtime which we expect to become the default
soon.

Before the error would be an unhandled top level exception without a
clear error message, now the exit code is set to `ExitCode.config` and a
more actionable error message about potential bad builder definitions is
printed.