Add storage and re-reporting of previous errors
Towards #910

The main benefit of the proposed `--rerun-failing-actions` is to be able
to see errors again from previous builds that have scrolled off screen.
The downside to rerunning is that it is a pattern encouraging retrying
non-hermetic actions. Instead we will store and re-print previous error
messages which more directly addresses the need.

This commit sketches out the interaction between the failure reporter
and the build - storage of errors to individual files will happen in a
followup.