Drop fail-on-severe flag
Closes #1097

We still have one blocking issue where we want to report errors from
previous builds - that implementation will need to change how errors
flow through the build (they need to be kept as Strings rather than a
boolean) so it will be easier to implement without this flag adding
noise.

Keep the argupment parsing to not break some CLI usages, but ignore it.