Add reset API to Resolvers
Towards #1584

We were already counting on the build system resetting information from
the AnalysisContext and we were doing so in different ways.
`build_runner` was using an explicit type for it's resolvers and
resetting through that interface, `bazel_codegen` was making assumptions
about the use of the Analyzer and resetting global state.

This makes the resetting explicit and adds an interface for it to
`Resolvers` so that the build systems can make fewer assumptions about
the particulars.