Add storage and re-reporting of previous errors (#1581)
Closes #910

The main benefit of the proposed `--rerun-failing-actions` is to be able
to see errors again from previous builds that have scrolled off screen.
The downside to rerunning is that it is a pattern encouraging retrying
non-hermetic actions. Instead we will store and re-print previous error
messages which more directly addresses the need.