Add a `ModuleLibrary` class
Towards #1644

Explicitly separate out the step of parsing a Dart source file and
classifying directives from the other work happening in the meta module.
The `ModuleLibrary` class is essentially the `_AssetNode` class from
before, except it is also responsible for determining which files should
be skipped and which are entrypoints rather than that responsibilty
being split with `_computeModules`.

"internal" and "external" dependencies are also not tracked separately
since it is applied as a filter only for the strongly connected
components call.

Eventually we will add in a capability to store a `ModuleLibrary` which
will let us write it as a separate Asset and then restore it to
compute meta modules.

One minor change in behavior is that we no longer throw if a "part"
directive reaches across packages or top-level directories. This should
be resolved if there are package level cycles by the next iteration of
strongly connected components.