Filter by non-hidden applied builders (#1653)
Closes #1578

Previously we filtered out any builders on non-root packages which would
attempt to write to source. Now we also filter out builders which write
to cache but want to apply some builder that would write to source,
under the assumption that if a builder can't be applied it isn't worth
doing the hidden work.

Add a test for the existing filtering, as well as the new filtering.