Add a bounded lru cache for files (#1660)
For now I have set the following defaults, which we can make configurable in the future:

~512MB per cache (one for bytes and one for strings)
1MB max per file, anything larger than that is never cached
I also fixed an issue where we actually were using the Encoding as a key before but it doesn't implement hashCode or == so it could have exploded in size since we had separate caches per encoding instance. It now uses the name of the Encoding, which should be pretty reasonable.

Note that sizes are based on the file size on disk and not necessarily how much memory is actually consumed by the vm representation of the objects.