Filter out non-lib assets in deps
Fixes #1740

We use a default glob of `lib/**` where there is no explicit target, but
honor exact globs when there is a target. We can nonly allow non-lib
sources in targets for the root package, so they are now filtered out
for deps.

Add a second package to the test setup for the `BuildDefinition` and add
a regression test for this case.