Drop DirectoryWatcherFactory typedef
Towards #1086

It doesn't hold it's weight - the Function type is not much longer, the
variable and argument names are already descriptive, and it's more clear
to see the definition inline than to have an indirection to see the
argument type.

The most important place to change this is `build.dart` where it is used
in a public API but the typedef itself is not public.