Log an actionable warning for missing config
Fixes #1841

Since there is a new place where we might log messages which can occur
before the `BuildOptions` can be constructed, add a `LogEnvironment`
which can be set up earlier.

Log an actionable warning when we fail to read the `buid.key.yaml` file
and throw a `CannotBuildException` which will bail out of the build and
exit with `ExitCode.config`.