Improve error message when build script is invalid
See #1900

Add explicit error handling in the places where we can identify either
syntactically or semantically invalid code.

If the script fails to parse when attempting to format the output
pointing to the error isn't very useful, since at that point the script
is all on one line - for this case indicate that the script cannot be
parsed.

If the script fails to snapshot we can print the stderr from the dart
process which should include details of the problem.

In both cases we assume that the problem stems from bad config and so we
exit with cod 78.

Also fix some bugs with the test utilities for integration tests using
`pub run build_runner` instead of running with a simulated manual build
script:

- Add missing dependencies for the pubspecs in `packageWithBuilders`.
- Create the directories for dependencies when creating with `package`.
- In pubspecs write both `dependencies` and `dependency_overrides` since
  overrides are not taken into account for the non-root package.