Prepare for breaking change to UriResolver.
Analyzer version 0.34.0 contains a breaking change to UriResolver: the
addition of a `clearCache` method - see
https://dart-review.googlesource.com/c/sdk/+/85726/1/pkg/analyzer/lib/src/generated/source.dart.

Even though we are not ready to upgrade to analyzer 0.34.0 yet, we can
prepare for the change by adding implementations of the method.  It's
not necessary for the implementations to do anything, since our
UriResolver implementations don't cache any URI resolution.