Include part files in build script update checks
Fixes #1912

Iterate over all declarations and pull out the source URIs for each.
This prevents the issue of only including the root of each library as an
asset used by the build script.

If a part file has no declarations it may be skipped, however an edit to
such a part file should have no impact on the build until an edit is
also made to another file to use the new declaration.