Allow PackageAssetReader without package as cwd
We can't rely on a pub layout for all packages because when running
tests under a `build_runner` generated output directory we omit the
`lib/` directory for non-root packages. It should be a safe assumption,
however, that the root package uses a pub layout whether we're running
from the actual package, a `build_runner` output, or a bazel test
environment.