Allow PackageAssetReader without package as cwd (#1939)
In some environments where we run tests we can't guarantee that `cwd` is the root package's directory. In environments other than `build_runner test` we do expect that the root directory will have a pub layout with a `lib` directory.

Check for a pub layout for the root package, and if it exists use it. Otherwise fallback to the `cwd` assumption.