Use xor to compute combined input hashes instead of md5 (#1961)
This should still be unique enough for our purposes (we don't use these digests in hash maps), and it allows us to use unordered sets which are more efficient, as well as being very fast on its own.

On angular gallery for edits to the material button example I saw incremental build time improvements of about 10%.

This will also allow us to add parallelism to the computeCombinedDigest function in a followup PR, without having to retain ordering.

There is one known potential bug this would introduce, which is if you were to swap the contents of two files the resulting combined digest would not change. In order to resolve that issue we include the AssetId.hashCode in the digest in the default AssetReader.digest implementation, and add a comment that custom implementations should include the asset id in their digest as well.