Explicit check for overwriting input extension (#2033)
Fixes #1715

When attempting to replicate the behavior of a Transformer that modified
it's input a user may try to configure a Builder that has the same
output extension as the input. Look for this case explicitly and emit a
specific error.

This prevents an infinite loop attempting to construct an AssetGraph.