Drop some unnecessary typedefs
Closes #1086

Remove some typedefs that are not saving enough space or used often
enough to justify a unique name.

Keep typedefs that are giving a significantly shorter or more clear name
to a function type.

Deprecate the `OnDelete` typedef in `build_runner_core` since it is
public, even though it's not used anywhere removing it would technically
be a breaking change.