Warn instead of create an output without assets
Fixes #1297

Since we don't know which assets will exist (only those which may exist)
before the build, save this check until we are outputting the directory.
The build that already happen may be a waste - but it will get reused if
the command is corrected to specify an output from a root that does have
assets in it.