Simplify logging newline handling
The extra check for a leading newline was added as a hack to get a
specific behavior around enforcing certain messages to not get cleared
when starting a new build in watch mode. We've since added a line of
dashes and are using trailing newlines to avoid erasing these messages,
so the extra logic is no longer necessary.