Don't allow dart:isolate on DDC (#2065)
We're pulling in the `dart.library.isolate` import choices and sending
them to DDC to compile, but then it's following the default import and
not finding a file. Since we ignore the `uri_does_not_exist` error the
actual error reported confusingly refers to a `part-of` directive that
does not actually exist in the file.