Allow running builds with fake `test` package
Fixes #2059

If there is a path or git dependency on some packge that _isn't_ the
official test package, but happens to be named `test`, it may not have a
version. That case should be detected later on when we look for
`build_test` which depends on the pub version of test (unless a
dependency override is used, in which case unexpected things may
happen).