Add output to DefaultBuildTarget (#2090)
The pattern we use for internal tools is to provide our own implementation of a `BuildTarget` and pass in the custom serializers to the Build Daemon client. Internal tools do not use the `DefaultBuildTarget` thus we can safely modify this implementation. As a result, the `DefaultBuildTarget` will refer to the `build_runner` implementation. This prevents us from needing yet another package as we don't want this logic in `build_runner_core` or `build_runner`.

Towards https://github.com/dart-lang/build/issues/2080