Fix exponential blowup in the kernel builder impl (#2119)
Specifically the problem this logic is solving is that any dependency which is provided as sources instead of kernel is essentially a "poison pill" node - all parents of those modules must also be provided as sources and not dill files.

The previous implementation worked around this in a weird way, checking all transitive deps (of each transtive dep) for package cycles. This was both inefficient and relied upon knowledge of the module structure and how it deals with package cycles which wasn't ideal.

The new logic handles this while collecting the transitive deps initially - it tracks module parents and any time it finds a module that has to be provided via source it marks all transitive parents as also being provided by source. It also bails early if the parent is already marked, preventing exponential behavior during marking.