Resolve graph before crawling parents (#2121)
Split out the transitive source and kernel dependency categorization out
into separate steps instead of mixing them as we crawl the graph. This
makes the logic a bit easier to follow since there are distinct phases
for crawling in each direction (dependencies vs parents). This might be
a slight performance improvement since `crawlAsync` is implemented with
`Future.wait` instead of crawling serially.