Update to new builder key format in tests
After the update to `build_config` version `0.3.2` the keys that we see
in the code no longer match what they were before. When we have
expectations against builders keys they need to use `:` as the
separator.

Manual calls to `apply` also need to use the new format since this is
what gets generated in the build script.

Also update the uses which mimic `build.yaml` files even though it isn't
strictly required.