Use the platform name for SDK dill digest
Prevents a potential bug where we fail to invalidate stored state in the
incremental compiler. This isn't an issue today since the 'vm' platform
is never able to use the incremental compiler due to passing
`--input-linked` arguments.

In the long term we'll want to consider using separate workers, or
setting up the kernel worker to be able to avoid a full invalidation
when hopping between platforms.