Cleanup Future<Null> to Future<void>
`Future<void>` is the recommended way to express what we want, clean up
the cases that were introduced before it was possible to write it this
way.

Some of these are technically breaking but are known to have no negative
impact on the build systems that use those code paths.

The interfaces implemented by users - Builder and Resource - are left as
is since it could be breaking to change them.