Allow passing .json file to graph_inspector
This is a breaking change in the command line parsing, but the tool is
already considered to not have a stable API.

Add and argument parser and mutually exclusive `--graph-file` and
`--build-script` arguments. Use a default for `--build-script` of
`.dart_tool/build/entrypoint/build.dart` since that's the typical case.

The `--graph-file` argument makes it possible to inspect graphs that
were produced on other machines etc.