Add clean command to runner binary (#2179)
Since the clean process does not need instantiated Builders it can run
in the outer binary rather than from the generated script. This should
make it faster. We also want to support `clean` for custom build scripts
so keep the implementation in `lib/` but add a separately configured
command from the outer binary as well.

- Move the non build script commands to a `bin/src/commands` directory.
- Make a list of "local" commands.
- Start logging before running local commands, explicitly disable
  logging in the one command which must not have any other stdout.