Report null assetId during deserialization (#2187)
Towards #1804
Followup on #2181

In a reported case where we were able to inspect the serialized asset
graph the problem exhibited is that a `null` was tracked through
`outputs`, rather than an ID which points to a missing node. Catch that
case specifically as well because the error message can be more direct.