Drop hardcoded list of known logger names
Now that there is a split with some of these logger names coming from
`build_runner_core` it's not as easy to roll out changes which introduce
new logger names. The only cases where we do want to show the logger
name are the ones where we are running actions and will always fit the
pattern of `'$builderLabel on $input'`.